GHI-364 Refactoring: extract in separate file
